+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Features Added
 
+- Added `UseProbeRequest` option for `ManagedIdentityCredential`.
+- By default, `ManagedIdentityCredential` does not send a probe request, unless it is a part of credential chain in `DefaultAzureCredential`.
+- When `AZURE_TOKEN_CREDENTIALS` environment variable is configured to `ManagedIdentityCredential`, the `DefaultAzureCredential` does not issue a probe request and performs retries with exponential backoff.
+
 ### Breaking Changes
 
 ### Bugs Fixed